separated lexer into its own module
