+++ src/xml/de/mod.rs
@@ -1,17 +1,17 @@
-#![deny(unused_must_use)]
-use super::error::*;
-use super::error::ErrorCode::*;
+#![deny(unused_must_use, unused_imports)]
+use xml::error::*;
+use xml::error::ErrorCode::*;
 use de;
 
-use std::iter::Peekable;
-
-use std::str;
+mod lexer;
+use self::lexer::Lexical::*;
+pub use self::lexer::LexerError;
 
 macro_rules! expect {
     ($sel:expr, $pat:pat, $err:expr) => {{
         match try!($sel.bump()) {
             $pat => {},
-            _ => return Err($sel.rdr.expected($err)),
+            _ => return Err($sel.expected($err)),
         }
     }}
 }
@@ -27,338 +27,22 @@ macro_rules! is_val {
     ($sel:expr, $i:ident, $err:expr) => {{
         match try!($sel.ch()) {
             $i(x) => x,
-            _ => return Err($sel.rdr.expected($err)),
+            _ => return Err($sel.expected($err)),
         }
     }}
 }
 
-#[derive(Debug, Copy, PartialEq, Clone)]
-pub enum Lexical<'a> {
-    StartTagClose,
-
-    Text(&'a [u8]),
-
-    StartTagName(&'a [u8]),
-
-    AttributeName(&'a [u8]),
-    AttributeValue(&'a [u8]),
-
-    EmptyElementEnd(&'a [u8]),
-
-    EndTagName(&'a [u8]),
-    StartOfFile,
-    EndOfFile,
-}
-
-#[derive(PartialEq)]
-enum InternalLexical {
-    StartTagClose,
-
-    Text,
-
-    StartTagName,
-
-    AttributeName,
-    AttributeValue,
-
-    EmptyElementEnd,
-
-    EndTagName,
-    StartOfFile,
-    EndOfFile,
-}
-
-enum LexerState {
-    Start,
-    AttributeName,
-    AttributeValue,
-    Tag,
-}
-
-struct LineColIterator<Iter: Iterator<Item=u8>> {
-    rdr: Peekable<Iter>,
-    line: usize,
-    col: usize,
-}
-
-impl<Iter: Iterator<Item=u8>> LineColIterator<Iter> {
-    fn peek(&mut self) -> Option<u8> {
-        self.rdr.peek().map(|&c| c)
-    }
-
-    fn expected(&self, reason: &'static str) -> Error {
-        self.error(Expected(reason))
-    }
-
-    fn error(&self, reason: ErrorCode) -> Error {
-        Error::SyntaxError(reason, self.line, self.col)
-    }
-
-    fn lexer_error(&self, reason: LexerError) -> Error {
-        Error::SyntaxError(LexingError(reason), self.line, self.col)
-    }
-
-    fn from_utf8<'a>(&self, txt: &'a[u8]) -> Result<&'a str, Error> {
-        let txt = str::from_utf8(txt);
-        txt.or(Err(Error::SyntaxError(NotUtf8, self.line, self.col)))
-    }
-}
-
-impl<Iter: Iterator<Item=u8>> Iterator for LineColIterator<Iter> {
-    type Item = u8;
-    fn next(&mut self) -> Option<u8> {
-        match self.rdr.next() {
-            None => None,
-            Some(b'\n') => {
-                self.line += 1;
-                self.col = 1;
-                print!(" -> \\n");
-                Some(b'\n')
-            },
-            Some(c) => {
-                print!(" -> {:?}", c as char);
-                self.col += 1;
-                Some(c)
-            },
-        }
-    }
-}
-
-struct XmlIterator<Iter: Iterator<Item=u8>> {
-    rdr: LineColIterator<Iter>,
-    buf: Vec<u8>,
-    stash: Vec<u8>,
-    state: LexerState,
-    ch: InternalLexical,
-}
-
-impl<Iter> XmlIterator<Iter>
-    where Iter: Iterator<Item=u8>,
-{
-    #[inline]
-    pub fn new(rdr: Iter) -> XmlIterator<Iter> {
-        XmlIterator {
-            rdr: LineColIterator {
-                rdr: rdr.peekable(),
-                line: 1,
-                col: 1,
-            },
-            buf: Vec::with_capacity(128),
-            stash: Vec::new(),
-            state: LexerState::Start,
-            ch: InternalLexical::StartOfFile,
-        }
-    }
-
-    fn stash(&mut self) {
-        use std::mem::swap;
-        swap(&mut self.buf, &mut self.stash);
-    }
-
-    fn stash_view(&self) -> &[u8] {
-        &self.stash
-    }
-
-    fn peek_char(&mut self) -> Result<u8, LexerError> {
-        self.rdr.peek().ok_or(LexerError::EOF)
-    }
-
-    fn next_char(&mut self) -> Result<u8, LexerError> {
-        self.rdr.next().ok_or(LexerError::EOF)
-    }
-
-    fn end(&mut self) -> Result<(), LexerError> {
-        while let Some(c) = self.rdr.next() {
-            if !b" \n\t\r".contains(&c) {
-                return Err(LexerError::ExpectedEOF);
-            }
-        }
-        Ok(())
-    }
-
-    fn decode(&mut self) -> Result<InternalLexical, LexerError> {
-        self.buf.clear();
-        match self.state {
-            LexerState::Start => self.decode_normal(),
-            LexerState::Tag => self.decode_tag(),
-            LexerState::AttributeName => self.decode_attr_name(),
-            LexerState::AttributeValue => self.decode_attr_val(),
-        }
-    }
-
-    fn decode_attr_val(&mut self) -> Result<InternalLexical, LexerError> {
-        let quot = self.rdr.find(|&ch| ch == b'\'' || ch == b'"');
-        let quot = try!(quot.ok_or(LexerError::EOF));
-        self.buf.extend(self.rdr.by_ref().take_while(|&ch| ch != quot));
-        // hack to detect EOF in take_while
-        try!(self.peek_char());
-        self.state = LexerState::AttributeName;
-        Ok(InternalLexical::AttributeValue)
-    }
-
-    fn decode_attr_name(&mut self) -> Result<InternalLexical, LexerError> {
-        use self::InternalLexical::*;
-        use self::LexerError::*;
-        loop {
-            return match try!(self.next_char()) {
-                b'/' => match try!(self.next_char()) {
-                    b'>' => {
-                        self.state = LexerState::Start;
-                        Ok(EmptyElementEnd)
-                    },
-                    _ => Err(ExpectedLT),
-                },
-                b'>' => {
-                    self.state = LexerState::Start;
-                    Ok(StartTagClose)
-                },
-                b'\n' | b'\r' | b'\t' | b' ' => continue,
-                c => {
-                    self.buf.push(c);
-                    break;
-                },
-            }
-        }
-        for c in &mut self.rdr {
-            if c == b'=' {
-                assert!(!self.buf.is_empty());
-                self.state = LexerState::AttributeValue;
-                return Ok(AttributeName);
-            }
-            if b"\n\r\t ".contains(&c) {
-                break;
-            }
-            self.buf.push(c);
-        }
-        for c in &mut self.rdr {
-            if c == b'=' {
-                assert!(!self.buf.is_empty());
-                self.state = LexerState::AttributeValue;
-                return Ok(AttributeName);
-            }
-            if !b"\n\r\t ".contains(&c) {
-                return Err(ExpectedEq);
-            }
-        }
-        Err(EOF)
-    }
-    fn decode_tag(&mut self) -> Result<InternalLexical, LexerError> {
-        use self::InternalLexical::*;
-        for c in &mut self.rdr {
-            if c == b'>' {
-                self.state = LexerState::Start;
-                return Ok(EndTagName)
-            }
-            self.buf.push(c);
-        }
-        Err(LexerError::EOF)
-    }
-
-    fn decode_tag_name(&mut self) -> Result<InternalLexical, LexerError> {
-        use self::InternalLexical::*;
-        loop {
-            match try!(self.peek_char()) {
-                b'\n' | b'\r' | b'\t' | b' ' | b'/' | b'>' => {
-                    debug_assert!(!self.buf.is_empty());
-                    self.state = LexerState::AttributeName;
-                    return Ok(StartTagName);
-                },
-                c => {
-                    self.buf.push(c);
-                    self.rdr.next();
-                }
-            }
-        }
-    }
-
-    fn decode_normal(&mut self) -> Result<InternalLexical, LexerError> {
-        use self::InternalLexical::*;
-        match self.rdr.next() {
-            Some(b'<') => match try!(self.next_char()) {
-                b'/' => {
-                    self.state = LexerState::Tag;
-                    Ok(Text)
-                },
-                b'!' => unimplemented!(),
-                b'?' => unimplemented!(),
-                c => {
-                    if self.buf.iter().all(|&c| b" \t\n\r".contains(&c)) {
-                        self.buf.clear();
-                        self.buf.push(c);
-                        self.decode_tag_name()
-                    } else {
-                        Err(LexerError::MixedElementsAndText)
-                    }
-                }
-            },
-            // error: not in this state
-            Some(c) if b"'\"=>/".contains(&c) => unimplemented!(),
-            Some(b'&') => unimplemented!(),
-            Some(c) => {
-                self.buf.push(c);
-                self.decode_normal()
-            },
-            None => Ok(EndOfFile),
-        }
-    }
-
-    fn ch(&self) -> Result<Lexical, Error> {
-        Ok(match self.ch {
-            InternalLexical::StartTagClose =>
-                Lexical::StartTagClose,
-            InternalLexical::Text =>
-                Lexical::Text(&self.buf),
-            InternalLexical::StartTagName =>
-                Lexical::StartTagName(&self.buf),
-            InternalLexical::AttributeName =>
-                Lexical::AttributeName(&self.buf),
-            InternalLexical::AttributeValue =>
-                Lexical::AttributeValue(&self.buf),
-            InternalLexical::EmptyElementEnd =>
-                Lexical::EmptyElementEnd(&self.buf),
-            InternalLexical::EndTagName =>
-                Lexical::EndTagName(&self.buf),
-            InternalLexical::StartOfFile =>
-                Lexical::StartOfFile,
-            InternalLexical::EndOfFile =>
-                Lexical::EndOfFile,
-        })
-    }
-
-    fn bump(&mut self) -> Result<Lexical, Error> {
-        print!("bump");
-        assert!(self.ch != InternalLexical::EndOfFile);
-        self.ch = match self.decode() {
-            Ok(ch) => ch,
-            Err(e) => return Err(self.rdr.lexer_error(e)),
-        };
-        println!(" -> {:?}", self.ch());
-        self.ch()
-    }
-}
-
-#[derive(Debug, Copy, PartialEq, Clone)]
-pub enum LexerError {
-    EOF,
-    ExpectedLT,
-    ExpectedQuotes,
-    Utf8,
-    MixedElementsAndText,
-    ExpectedEOF,
-    ExpectedEq,
-}
-
 pub struct Deserializer<Iter: Iterator<Item=u8>> {
-    rdr: XmlIterator<Iter>,
+    rdr: lexer::XmlIterator<Iter>,
 }
 
 pub struct InnerDeserializer<'a, Iter: Iterator<Item=u8> + 'a> (
-    &'a mut XmlIterator<Iter>, &'a mut bool
+    &'a mut lexer::XmlIterator<Iter>, &'a mut bool
 );
 
 impl<'a, Iter: Iterator<Item=u8> + 'a> InnerDeserializer<'a, Iter> {
     fn decode<T>(
-        xi: &mut XmlIterator<Iter>
+        xi: &mut lexer::XmlIterator<Iter>
     ) -> (bool, Result<T, Error>)
     where T : de::Deserialize
     {
@@ -379,13 +63,12 @@ where Iter: Iterator<Item=u8>,
     fn visit<V>(&mut self, mut visitor: V) -> Result<V::Value, Error>
         where V: de::Visitor,
     {
-        use self::Lexical::*;
         println!("InnerDeserializer::visit");
         match try!(self.0.ch()) {
             StartTagClose => {
                 match {
                     let v = expect_val!(self.0, Text, "text");
-                    let v = try!(self.0.rdr.from_utf8(v));
+                    let v = try!(self.0.from_utf8(v));
                     visitor.visit_str(v)
                 } { // try! is broken sometimes
                     Ok(v) => {
@@ -396,7 +79,7 @@ where Iter: Iterator<Item=u8>,
                 }
             },
             EmptyElementEnd(_) => visitor.visit_unit(),
-            _ => Err(self.0.rdr.expected("start tag close")),
+            _ => Err(self.0.expected("start tag close")),
         }
     }
 
@@ -500,21 +183,17 @@ impl<Iter> Deserializer<Iter>
     #[inline]
     pub fn new(rdr: Iter) -> Deserializer<Iter> {
         Deserializer {
-            rdr: XmlIterator::new(rdr),
+            rdr: lexer::XmlIterator::new(rdr),
         }
     }
 
-    fn ch(&self) -> Result<Lexical, Error> {
+    fn ch(&self) -> Result<lexer::Lexical, Error> {
         self.rdr.ch()
-    }
-
-    fn error(&self, reason: ErrorCode) -> Error {
-        Error::SyntaxError(reason, self.rdr.rdr.line, self.rdr.rdr.col)
     }
 
     fn end(&mut self) -> Result<(), Error> {
         match try!(self.ch()) {
-            Lexical::EndOfFile => Ok(()),
+            EndOfFile => Ok(()),
             _ => unimplemented!(),
         }
     }
@@ -530,7 +209,6 @@ impl<Iter> de::Deserializer for Deserializer<Iter>
     fn visit<V>(&mut self, visitor: V) -> Result<V::Value, Error>
         where V: de::Visitor,
     {
-        use self::Lexical::*;
         println!("Deserializer::visit");
         expect!(self.rdr, StartTagName(_), "start tag name");
         try!(self.rdr.bump());
@@ -540,7 +218,7 @@ impl<Iter> de::Deserializer for Deserializer<Iter>
         match try!(self.rdr.ch()) {
             EndTagName(_) => {},
             EmptyElementEnd(_) => {},
-            _ => return Err(self.rdr.rdr.expected("end tag")),
+            _ => return Err(self.rdr.expected("end tag")),
         }
         expect!(self.rdr, EndOfFile, "end of file");
         Ok(v)
@@ -571,7 +249,6 @@ impl<Iter> de::Deserializer for Deserializer<Iter>
     fn visit_map<V>(&mut self, visitor: V) -> Result<V::Value, Error>
         where V: de::Visitor,
     {
-        use self::Lexical::*;
         println!("Deserializer::visit_map");
         expect!(self.rdr, StartTagName(_), "start tag name"); // TODO: named map
         try!(self.rdr.bump());
@@ -580,7 +257,7 @@ impl<Iter> de::Deserializer for Deserializer<Iter>
         assert!(!*is_seq);
         match try!(self.ch()) {
             EndTagName(_) | EmptyElementEnd(_) => {},
-            _ => return Err(self.rdr.rdr.expected("end tag")),
+            _ => return Err(self.rdr.expected("end tag")),
         }
         expect!(self.rdr, EndOfFile, "end of file");
         Ok(v)
@@ -654,7 +331,7 @@ impl de::MapVisitor for EmptyMapVisitor {
 struct ContentVisitor<'a, Iter: 'a>
     where Iter: Iterator<Item=u8>,
 {
-    de: &'a mut XmlIterator<Iter>,
+    de: &'a mut lexer::XmlIterator<Iter>,
     state: ContentVisitorState,
 }
 
@@ -669,7 +346,7 @@ enum ContentVisitorState {
 impl <'a, Iter> ContentVisitor<'a, Iter>
     where Iter: Iterator<Item=u8>,
 {
-    fn new_attr(de: &'a mut XmlIterator<Iter>) -> Self {
+    fn new_attr(de: &'a mut lexer::XmlIterator<Iter>) -> Self {
         ContentVisitor {
             de: de,
             state: ContentVisitorState::Attribute,
@@ -685,14 +362,13 @@ impl<'a, Iter> de::MapVisitor for ContentVisitor<'a, Iter>
     fn visit_key<K>(&mut self) -> Result<Option<K>, Error>
         where K: de::Deserialize,
     {
-        use self::Lexical::*;
         use self::ContentVisitorState::*;
         println!("visit_key: {:?}", (&self.state, try!(self.de.ch())));
         match match (&self.state, try!(self.de.ch())) {
             (&Attribute, EmptyElementEnd(_)) => return Ok(None),
             (&Attribute, StartTagClose) => 0,
-            (&Attribute, AttributeName(n)) => return Ok(Some(try!(KeyDeserializer::visit(try!(self.de.rdr.from_utf8(n)))))),
-            (&Element, StartTagName(n)) => return Ok(Some(try!(KeyDeserializer::visit(try!(self.de.rdr.from_utf8(n)))))),
+            (&Attribute, AttributeName(n)) => return Ok(Some(try!(KeyDeserializer::visit(try!(self.de.from_utf8(n)))))),
+            (&Element, StartTagName(n)) => return Ok(Some(try!(KeyDeserializer::visit(try!(self.de.from_utf8(n)))))),
             (&Inner, Text(_)) => 1,
             (&Inner, _) => 4,
             (&Value, EndTagName(_)) => return Ok(None),
@@ -744,14 +420,13 @@ impl<'a, Iter> de::MapVisitor for ContentVisitor<'a, Iter>
     fn visit_value<V>(&mut self) -> Result<V, Error>
         where V: de::Deserialize,
     {
-        use self::Lexical::*;
         use self::ContentVisitorState::*;
         println!("visit_value: {:?}", &self.state);
         match self.state {
             Attribute => {
                 let v = {
                     let v = expect_val!(self.de, AttributeValue, "attribute value");
-                    let v = try!(self.de.rdr.from_utf8(v));
+                    let v = try!(self.de.from_utf8(v));
                     try!(KeyDeserializer::visit(v))
                 };
                 try!(self.de.bump());
@@ -766,7 +441,7 @@ impl<'a, Iter> de::MapVisitor for ContentVisitor<'a, Iter>
                     match try!(self.de.ch()) {
                         EmptyElementEnd(_) => {},
                         EndTagName(_) => {},
-                        _ => return Err(self.de.rdr.expected("tag close")),
+                        _ => return Err(self.de.expected("tag close")),
                     }
                     try!(self.de.bump());
                 }
@@ -775,7 +450,7 @@ impl<'a, Iter> de::MapVisitor for ContentVisitor<'a, Iter>
             Value => {
                 let v = {
                     let v = is_val!(self.de, Text, "text");
-                    let v = try!(self.de.rdr.from_utf8(v));
+                    let v = try!(self.de.from_utf8(v));
                     try!(KeyDeserializer::visit(v))
                 };
                 try!(self.de.bump());
@@ -800,14 +475,14 @@ impl<'a, Iter> de::MapVisitor for ContentVisitor<'a, Iter>
 }
 
 struct SeqVisitor<'a, Iter: 'a + Iterator<Item=u8>> {
-    de: &'a mut XmlIterator<Iter>,
+    de: &'a mut lexer::XmlIterator<Iter>,
     done: bool,
 }
 
 impl<'a, Iter> SeqVisitor<'a, Iter>
     where Iter: Iterator<Item=u8>,
 {
-    fn new(de: &'a mut XmlIterator<Iter>) -> Self {
+    fn new(de: &'a mut lexer::XmlIterator<Iter>) -> Self {
         SeqVisitor {
             de: de,
             done: false,
@@ -823,7 +498,6 @@ impl<'a, Iter> de::SeqVisitor for SeqVisitor<'a, Iter>
     fn visit<T>(&mut self) -> Result<Option<T>, Error>
         where T: de::Deserialize,
     {
-        use self::Lexical::*;
         println!("SeqVisitor::visit: {:?}", (self.done, self.de.ch()));
         if self.done {
             return Ok(None);
@@ -831,7 +505,7 @@ impl<'a, Iter> de::SeqVisitor for SeqVisitor<'a, Iter>
         let (is_seq, v) = InnerDeserializer::decode(&mut self.de);
         let v = try!(v);
         if is_seq {
-            return Err(self.de.rdr.error(XmlDoesntSupportSeqofSeq));
+            return Err(self.de.error(XmlDoesntSupportSeqofSeq));
         }
         is_val!(self.de, EndTagName, "end tag");
         self.de.stash();
